requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
requests==2.31.0
feedparser==6.0.10
jsonpath-ng==1.6.1
Werkzeug==3.0.1